Divide genres list from method
