+++ public/env.js
@@ -1,12 +1,12 @@
 window.env = {
-  "TRENDING": "https://api.themoviedb.org/3/trending/movie/week",
-  "DETAILS": "https://api.themoviedb.org/3/movie/",
-  "LATEST": "https://api.themoviedb.org/3/movie/latest",
-  "NOW_PLAYING": "https://api.themoviedb.org/3/movie/now_playing",
-  "POPULAR": "https://api.themoviedb.org/3/movie/popular",
-  "TOP_RATED": "https://api.themoviedb.org/3/movie/top_rated",
-  "UPCOMING": "https://api.themoviedb.org/3/movie/upcoming",
-  "API_KEY": "42cea6b932576c057f794e0128e498c2",
-  "IMG_URL": "https://image.tmdb.org/t/p/original",
-  "POSTER_URL": "https://image.tmdb.org/t/p/w342"
+  TRENDING: 'https://api.themoviedb.org/3/trending/movie/week',
+  DETAILS: 'https://api.themoviedb.org/3/movie/',
+  LATEST: 'https://api.themoviedb.org/3/movie/latest',
+  NOW_PLAYING: 'https://api.themoviedb.org/3/movie/now_playing',
+  POPULAR: 'https://api.themoviedb.org/3/movie/popular',
+  TOP_RATED: 'https://api.themoviedb.org/3/movie/top_rated',
+  UPCOMING: 'https://api.themoviedb.org/3/movie/upcoming',
+  API_KEY: '42cea6b932576c057f794e0128e498c2',
+  IMG_URL: 'https://image.tmdb.org/t/p/original',
+  POSTER_URL: 'https://image.tmdb.org/t/p/w342',
 };
