+++ src/styles/bighero/BigHero.scss
@@ -91,6 +91,4 @@
       }
     }
   }
-
-  
 }
